requirements.txt
asgiref==3.8.1
Django==4.2.8
django-bootstrap5==25.1
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
